Solver basics in dirty CLI
This was typed out in one go, but seems to work. It could use some improvements, though. Move the solving to a separate class, maybe not static.

Instead of reading from a file in code, allow passing a filename, or piping in puzzle input, and accept input from the command line. 

So much to do, so little time. :--|